split `traits.rs` to smaller files too
